Add progress for Step4
