+++ app/containers/Main.js
@@ -1,15 +1,33 @@
 var React = require('react');
-var container = require('../styles').container;
+var ZipCodeContainer = require('./ZipCodeContainer');
 
-var HelloWorld = React.createClass({
+var styles = {
+  container: {
+    width: '100%',
+    height: '92%'
+  },
+  header :{
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    background: 'rgba(252, 90, 44, 0.89)',
+    color: '#fff',
+    padding: 5,
+  }
+}
+
+var Main = React.createClass({
   render: function () {
     return (
-      <div style={container}>
-        <h1>Main.js Header</h1>
+      <div style={styles.container}>
+        <div style={styles.header}>
+          <h2 style={{margin: 0}}>Clever Title</h2>
+          <ZipCodeContainer direction='row' />
+        </div>
         {this.props.children}
       </div>
     )
   }
 })
 
-module.exports = HelloWorld;
+module.exports = Main;
