--- app/styles/index.js
@@ -1,8 +1,0 @@
-var styles = {
-  container: {
-    height: '100%',
-    width: '100%',
-  }
-};
-
-module.exports = styles;
